@@ -3,7 +3,7 @@
   "version": "6.0.2",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --port=4211",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -11,42 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.2.0",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "^8.0.1",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.1",
+    "@angular/compiler": "^8.0.1",
+    "@angular/core": "^8.0.1",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.0.1",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "^8.0.1",
+    "@angular/platform-browser-dynamic": "^8.0.1",
+    "@angular/router": "^8.0.1",
+    "core-js": "^3.1.3",
+    "rxjs": "^6.5.2",
+    "rxjs-compat": "^6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
-    "@angular-devkit/build-ng-packagr": "~0.8.0",
-    "@angular/cli": "~6.2.4",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "^0.800.3",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "~8.0.3",
+    "@angular/compiler-cli": "^8.0.1",
+    "@angular/language-service": "^6.1.10",
+    "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "@types/node": "^12.0.8",
+    "codelyzer": "^5.1.0",
+    "jasmine-core": "^3.4.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.1.0",
-    "protractor": "~5.4.0",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^5.1.0",
+    "protractor": "^5.4.2",
+    "rxjs-tslint": "^0.1.7",
     "ts-node": "~7.0.0",
-    "tsickle": ">=0.29.0",
+    "tsickle": "^0.35.0",
     "tslib": "^1.9.0",
-    "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "tslint": "^5.17.0",
+    "typescript": "^3.4.5",
+    "webpack": "^4.34.0"
   }
 }
