@@ -1,5 +1,5 @@
 import {NgModule, InjectionToken, Optional, Inject, isDevMode} from '@angular/core';
-import {DOCUMENT} from '@angular/platform-browser';
+import {DOCUMENT} from '@angular/common';
 import {CompatibilityModule} from '../compatibility/compatibility';
 
 
